+++ MyApplication/app/src/main/java/features/JuniorConnectPending.java
@@ -135,8 +135,7 @@ public class JuniorConnectPending extends Fragment {
                 String line = "";
                 while((line = bufferedReader.readLine()) != null)
                     sb.append(line + '\n');
-                getCards(sb.toString());
-                return "success";
+                return getCards(sb.toString());
             } catch (Exception e) {
                 e.printStackTrace();
                 return "error";
@@ -158,31 +157,39 @@ public class JuniorConnectPending extends Fragment {
             else if (result.equals("error")){
                 Toast.makeText(getContext(), "Unable to update! Check network connection", Toast.LENGTH_SHORT).show();
             }
+            else
+                Toast.makeText(getContext(), "Sorry! Unable to update!", Toast.LENGTH_SHORT).show();
             cardArrayAdapter.refresh();
         }
     }
-    private void getCards(String result){
-        JSONObject jObject= null;
-        try {
-            jObject = new JSONObject(result);
-            JSONArray jArray=jObject.getJSONArray("requests");
-            int len=jArray.length();
-            db.deletePendingJuniors();
-            for(int i=0; i<len;i++){
+    private String getCards(String result){
 
-                JSONObject object=jArray.getJSONObject(i);
-                JuniorModel model= new JuniorModel();
-                model.name=object.getString("name");
-                model.town=object.getString("hometown");
-                model.state=(new JSONObject(object.getString("state"))).getString("name");
-                model.branch=(new JSONObject(object.getString("branch"))).getString("name");
-                model.username=object.getString("username");
-                model.description=object.getString("description");
-                model.status="pending";
-                db.addJunior(model);
+        try {
+            JSONObject jObject = new JSONObject(result);
+            if ("success".equals(jObject.getString("status"))){
+                JSONArray jArray=jObject.getJSONArray("requests");
+                int len=jArray.length();
+                db.deletePendingJuniors();
+                for(int i=0; i<len;i++){
+
+                    JSONObject object=jArray.getJSONObject(i);
+                    JuniorModel model= new JuniorModel();
+                    model.name=object.getString("name");
+                    model.town=object.getString("hometown");
+                    model.state=(new JSONObject(object.getString("state"))).getString("name");
+                    model.branch=(new JSONObject(object.getString("branch"))).getString("name");
+                    model.username=object.getString("username");
+                    model.description=object.getString("description");
+                    model.status="pending";
+                    db.addJunior(model);
+                }
+                return "success";
             }
+            else
+                return "fail";
         } catch (JSONException e) {
             e.printStackTrace();
+            return "fail";
         }
     }
     public class JuniorCardArrayAdapter  extends ArrayAdapter<JuniorModel> {
